fix: Uncontrolled data security issue
